requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
audioop-lts==0.2.1
certifi==2024.12.14
charset-normalizer==3.4.1
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.1
py-cord==2.6.1
requests==2.32.3
urllib3==2.3.0
yarl==1.18.3
